Fixed read-only execute.
- We open the currently executing file and leave it open until the
  thread exits so that it can't be written to.
- Passing all but 1 of 76 tests (multi-oom).

+++ src/lib/kernel/vector.c
@@ -100,7 +100,7 @@ void vector_remove(struct vector *v, unsigned int pos)
     unsigned int i;
 
     ASSERT(v);
-    for (i = pos; i < (unsigned int) v->size; i++)
+    for (i = pos; i < v->size; i++)
         v->data[i] = v->data[i + 1];
     v->size--;
 }
